FEAT, FIX: added diffrax evolution example, fixed JAX config
With test/evolution_granad.py we have an example of how to use diffrax
for the time propagation. This promises speedups.

Additionally, since we don't depend on numba/pyfftw anymore, we can
safely use newer Python/JAX versions. This requires adjusting the
config for the float precision in the main code.