FEAT, DOC: added gradient descent example and simplified stacking
stack construction can now happen without specifying coulomb
interaction via an additional argument. if set to False, the coulomb
matrix is zeroed out.

also, we now have a quick function to do "quasi-nn" couplings. the
idea is that we construct a sequence of very narrow gaussians centered
at the distances we have in our structure.